Move and reuse TestHelper
